adding changes to Authorize.net Model
Basically just adding the items to the request, and changing the post
method from using the Zend array based request to the raw method.  The
reason the array method was incompatible is that authorize.net requires
line items to come in as multiple “x_line_item”, and the default array
method cannot support that.